So, I changed a few things. It's basically just a matter of classes, properties and packages organization, since I believe that it was a little bit messy.
After upgrading SpringBoot's version to 2+, a lot of annotations were
causing compilation error. I removed them and ajusted the classes to use
the correct way (so far).

Created the application.properties file to use in the test
(local.server.port).

Created the resource to be registered, since this is a JAX-RS way of
creating Rest services. I believe that this is not necessary,
but will keep it, since I don't know how far I can change things in this
test (I believe that we don't have to use that style, and should be
using MVC, but I won't refactor to that point, since, as I said, I don't
know how much I can change this code).